+++ Editor/Drawers/ColorToggleAttributeDrawer.cs
@@ -358,7 +358,7 @@ namespace SaintsField.Editor.Drawers
         protected override void OnValueChanged(SerializedProperty property, ISaintsAttribute saintsAttribute, int index,
             VisualElement container,
             FieldInfo info,
-            object parent, object newValue)
+            object parent, Action<object> onValueChangedCallback, object newValue)
         {
             UpdateToggleDisplay(property, saintsAttribute, index, container, info, parent);
         }