+++ Editor/Drawers/EnumFlagsAttributeDrawer.cs
@@ -709,7 +709,7 @@ namespace SaintsField.Editor.Drawers
         protected override void OnValueChanged(SerializedProperty property, ISaintsAttribute saintsAttribute, int index,
             VisualElement container,
             FieldInfo info,
-            object parent, object newValue)
+            object parent, Action<object> onValueChangedCallback, object newValue)
         {
             int newInt = (int)newValue;
             UpdateButtonDisplay(newInt, property, container);