+++ Samples~/Scripts/FieldInterfaceExample/FieldInterfaceMono.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-namespace SaintsField.Samples.Scripts.FieldInterfaceExample
-{
-    public class FieldInterfaceMono: MonoBehaviour
-    {
-        [FieldInterface(typeof(IMyInterface))] public SpriteRenderer interSr;
-    }
-}