+++ Samples~/Scripts/FieldInterfaceExample/IMyInterface.cs
@@ -1,6 +1,0 @@
-namespace SaintsField.Samples.Scripts.FieldInterfaceExample
-{
-    public interface IMyInterface
-    {
-    }
-}