+++ Samples~/Scripts/FieldInterfaceExample/MySubInter.cs
@@ -1,7 +1,0 @@
-﻿namespace SaintsField.Samples.Scripts.FieldInterfaceExample
-{
-    public class MySubInter: MyInter1
-    {
-
-    }
-}