+++ Samples~/Scripts/RequiredTypeExample/IMyInterface.cs
@@ -1,0 +1,6 @@
+namespace SaintsField.Samples.Scripts.RequiredTypeExample
+{
+    public interface IMyInterface
+    {
+    }
+}