+++ Samples~/Scripts/RequiredTypeExample/MyInter1.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace SaintsField.Samples.Scripts.FieldInterfaceExample
+namespace SaintsField.Samples.Scripts.RequiredTypeExample
 {
     public class MyInter1: MonoBehaviour, IMyInterface
     {