+++ Samples~/Scripts/RequiredTypeExample/MySubInter.cs
@@ -1,0 +1,6 @@
+﻿namespace SaintsField.Samples.Scripts.RequiredTypeExample
+{
+    public class MySubInter: MyInter1
+    {
+    }
+}