+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.2.1
+
+1.  Add `RequireType`. This attribute allow you to specify the required component or **interface** for a field.
+2.  `FieldType` now has a custom object picker window. By default Unity disallow to see an asset with a specific type because of "performance consideration".
+    Ironically, if you directly use a component type instead of a `GameObject`, Unity **WILL** show a correct asset picker.
+    As this can not be "hacked" or "tricked" to let Unity show the correct picker, the custom object picker will allow you to pick a object (e.g. a prefab)
+    with the component you want. 
+
 ## 2.2.0
 
 1.  Fix [Issue 8](https://github.com/TylerTemp/SaintsField/issues/8) that most attributes in `SaintsField` did NOT looking for the inherited parent target... This includes `PropRange`, `Min`, `Max`, `Dropdown`, `AdvancedDropdown`, `SpriteToggle`, `MaterialToggle`, `ColorToggle`, `RichLabel`, `Above/BelowRichLabel`, `InfoBox`, Buttons etc.
@@ -15,7 +23,7 @@ Special thanks to [ZeroUltra](https://github.com/ZeroUltra) to find this importa
     failed to display some fields.
 2.  Remove default `SaintsEditor` for example scene, so people who imports it (most likely when using `unitypackage`) won't accidentally
     get `SaintsEditor` enabled for the whole project.
-3.  Fix a bug that possibliy break `rate` in some situation.
+3.  Fix a bug that possibly break `rate` in some situation.
 4.  IMGUI: fix `SaintsEditor` display an empty `MonoScript` when the target is neither `MonoBehavior` nor `ScriptableObject`.
 
 ## 2.1.12